Update to latest spec and test cases
